pyproject.toml
[project]
name = "latita"
version = "0.1.0"
description = "Ephemeral libvirt/QEMU lab manager with capsules"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "questionary>=2.1",
  "pyyaml>=6.0",
]

[project.scripts]
latita = "latita.cli:main"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/latita"]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
